Module Dcr.Evaluation

val eval_expr : Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result

eval_expr expr expr_env evaluates the given expression expr in the environment expr_env.

parameter expr
The expression to evaluate.

parameter expr_env
The environment in which to evaluate the expression.

returns
The result of evaluating the expression.

val partial_eval_expr : Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result

partial_eval_expr expr expr_env partially evaluates the given expression expr in the environment expr_env, i.e. evaluates the expression as much as possible without knowing the values of all the variables. This is useful for evaluating expressions in the context of a specific environment or for optimizing expressions.

parameter expr
The expression to partially evaluate.

parameter expr_env
The environment in which to partially evaluate the expression.

returns
The result of partially evaluating the expression.