Module Dcr.Helper

val update_event : Ast.Syntax.event' Ast.Syntax.annotated -> Ast.Syntax.program -> Ast.Syntax.program

update_event event program updates the event in the program by replacing the old event with the new one.

parameter event
The new event to be updated.

parameter program
The program where the event is to be updated.

returns
The updated program.

val set_marking : ?included:bool -> ?pending:bool -> ?executed:bool -> ?value:Ast.Syntax.expr -> Ast.Syntax.event' Ast.Syntax.annotated -> (Ast.Syntax.event' Ast.Syntax.annotated, Ast.Error.detailed_error list) Stdlib.result

val get_marking : Ast.Syntax.event' Ast.Syntax.annotated -> Ast.Syntax.event_marking'

val is_included : Ast.Syntax.event' Ast.Syntax.annotated -> bool

val is_executed : Ast.Syntax.event' Ast.Syntax.annotated -> bool

val is_pending : Ast.Syntax.event' Ast.Syntax.annotated -> bool

val current_value : Ast.Syntax.event' Ast.Syntax.annotated -> Ast.Syntax.expr

val update_event_io : ?eval: (Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result) -> Ast.Syntax.event' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.event' Ast.Syntax.annotated, Ast.Error.detailed_error list) Stdlib.result

val update_event_info : ?eval: (Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result) -> Ast.Syntax.event' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.event' Ast.Syntax.annotated, Ast.Error.detailed_error list) Stdlib.result

val set_info : ?id:string -> ?label:string -> Ast.Syntax.event' Ast.Syntax.annotated -> (Ast.Syntax.event' Ast.Syntax.annotated, Ast.Error.detailed_error list) Stdlib.result

val update_event_env : Ast.Syntax.program -> Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env, 'a) Stdlib.result

val change_relation : string Ast.Syntax.annotated -> Ast.Syntax.event_id -> Ast.Syntax.relation' Ast.Syntax.annotated -> Ast.Syntax.relation' Ast.Syntax.annotated

val update_relation_guard : ?eval: (Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result) -> Ast.Syntax.relation' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.relation' Ast.Syntax.annotated, Ast.Error.detailed_error list) Stdlib.result

val update_expr_id : ?eval: (Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result) -> string -> string Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr' Ast.Syntax.annotated, Ast.Error.detailed_error list) Stdlib.result

val preprocess_program : ?expr_env:Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> ?event_env:Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env -> Ast.Syntax.program -> (Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env * Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env * Ast.Syntax.program, 'a) Stdlib.result

val preprocess_subprogram : ?expr_env:Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> ?event_env:Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.event list * Ast.Syntax.template_instance list * Ast.Syntax.relation list * Ast.Syntax.template_annotation' list) -> (Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env * Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env * (Ast.Syntax.event list * Ast.Syntax.template_instance list * Ast.Syntax.relation list * Ast.Syntax.template_annotation' list), 'a) Stdlib.result

val get_event : ?filter:(Ast.Syntax.event -> bool) -> Ast.Syntax.program -> Ast.Syntax.event option

get_event ?filter program returns the event with the identifier id in the program

val has_event : ?filter:(Ast.Syntax.event -> bool) -> Ast.Syntax.program -> bool

val find_event : string Ast.Syntax.annotated -> 'a Common.Env.env -> ('a, Ast.Error.detailed_error list) Stdlib.result

val find_all_events : ?filter:(Ast.Syntax.event -> bool) -> Ast.Syntax.program -> (Ast.Syntax.event list, 'a) Stdlib.result

val same_id : string -> Ast.Syntax.event' Ast.Syntax.annotated -> bool

val get_relation : ?filter:(Ast.Syntax.relation' Ast.Syntax.annotated -> bool) -> string -> Ast.Syntax.program -> Ast.Syntax.relation' Ast.Syntax.annotated option

val is_spawn : Ast.Syntax.relation' Ast.Syntax.annotated -> bool

val is_ctrl : Ast.Syntax.relation_type -> Ast.Syntax.relation' Ast.Syntax.annotated -> bool

val has_relation : ?filter:(Ast.Syntax.relation' Ast.Syntax.annotated -> bool) -> string -> Ast.Syntax.program -> bool

val find_all_relations : ?filter: (Ast.Syntax.relation' Ast.Syntax.annotated -> Ast.Syntax.event_id -> Ast.Syntax.event_id -> bool) -> Ast.Syntax.program -> (Ast.Syntax.relation' Ast.Syntax.annotated list, 'a) Stdlib.result

val is_event_present_on_relation : string Ast.Syntax.annotated -> Ast.Syntax.relation' Ast.Syntax.annotated -> bool

val append_subprograms : ('a list * 'b list * 'c list * 'd list) list -> ('a list * 'b list * 'c list * 'd list, 'e) Stdlib.result

val event_as_expr : Ast.Syntax.event' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated

val sort_events : Ast.Syntax.program -> (Ast.Syntax.program, 'a) Stdlib.result

val sort_relations : Ast.Syntax.program -> (Ast.Syntax.program, 'a) Stdlib.result

val sort_program : Ast.Syntax.program -> (Ast.Syntax.program, 'a) Stdlib.result

val count : int Stdlib.ref

val counter : unit -> string

val nanoid : ?length:int -> unit -> string

val fresh : ?alpha_rename:(unit -> string) -> string -> string

val fresh_event : ?alpha_rename:(unit -> string) -> Ast.Syntax.event' Ast.Syntax.annotated -> (Ast.Syntax.event' Ast.Syntax.annotated, Ast.Error.detailed_error list) Stdlib.result

val fresh_events : ?exclude:string list -> ?alpha_rename:(unit -> string) -> (Ast.Syntax.event list * Ast.Syntax.template_instance list * Ast.Syntax.relation list * Ast.Syntax.template_annotation' list) -> (Ast.Syntax.subprogram, Ast.Error.detailed_error list) Stdlib.result

val fresh_event_ids : ?exclude:Stdlib.String.t list -> ?alpha_rename:(unit -> string) -> Ast.Syntax.subprogram -> (Ast.Syntax.subprogram, Ast.Error.detailed_error list) Stdlib.result

val bind_events : f:(Ast.Syntax.event' Ast.Syntax.annotated -> 'a) -> Ast.Syntax.event' Ast.Syntax.annotated list -> 'a Common.Env.env -> ('a Common.Env.env, 'b) Stdlib.result

val update_ids : (string * string Ast.Syntax.annotated) list -> Ast.Syntax.subprogram -> (Ast.Syntax.subprogram, 'a) Stdlib.result

val eval_bool : ?eval: (Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result) -> Ast.Syntax.expr' Ast.Syntax.annotated -> Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> (bool, Ast.Error.detailed_error list) Stdlib.result