Module Dcr.Instantiation

val instantiate : ?expr_env:Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env -> ?event_env:Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env -> Ast.Syntax.program -> (Ast.Syntax.program * Ast.Syntax.event' Ast.Syntax.annotated Common.Env.env * Ast.Syntax.expr' Ast.Syntax.annotated Common.Env.env, Ast.Error.detailed_error list) Stdlib.result

instantiate ?expr_env ?event_env program instantiates the templates and the meta-programming operators present in program.

parameter expr_env
The environment of expressions to instantiate against.

parameter event_env
The environment of events to instantiate against.

parameter program
The program to instantiate.

returns
A result containing the instantiated program, updated event environment and updated expression environment, or a list of errors.