Module Dcr.Transition

val is_enabled : Ast.Syntax.event -> Ast.Syntax.program -> (Ast.Syntax.event Common.Env.env * Ast.Syntax.expr Common.Env.env) -> (bool, Ast.Error.detailed_error list) Stdlib.result

is_enabled event program (event_env, expr_env) checks if the event event is enabled in the program program.

parameter event
The event to check if it is enabled.

parameter program
The program to check if the event is enabled.

parameter event_env
The environment with the events.

parameter expr_env
The environment with the expressions.

returns
result with a boolean indicating if the event is enabled.

val propagate_effects : Ast.Syntax.event -> (Ast.Syntax.event Common.Env.env * Ast.Syntax.expr Common.Env.env) -> Ast.Syntax.program -> (Ast.Syntax.program * Ast.Syntax.event Common.Env.env * Ast.Syntax.expr Common.Env.env, Ast.Error.detailed_error list) Stdlib.result

propagate_effects event (event_env, expr_env) program propagates the effects of the event event on the program program.

parameter event
The event to propagate the effects.

parameter event_env
The environment with the events.

parameter expr_env
The environment with the expressions.

parameter program
The program to propagate the effects.

returns
result with the program with the effects propagated.