Module Parsing.Lex_and_parse

Lexing and Parsing

This module provides functions to parse a lexbuf into a valid program or expression.

val parse_program : ?filename:string -> Ast.Syntax.Lexing.lexbuf -> (Ast.Syntax.program, Ast.Error.detailed_error list) Stdlib.result

parse_program ?filename lexbuf parses the lexbuf as a program and returns the result. This function sets the pos_fname field in the lexbuf to the provided filename. If an error occurs, it returns the error.

parameter filename
The name of the file being parsed.

parameter lexbuf
The lexbuf to be parsed.

returns
The result of the parsing as result.

val parse_expression : ?filename:string -> Ast.Syntax.Lexing.lexbuf -> (Ast.Syntax.expr, Ast.Error.detailed_error list) Stdlib.result

parse_expression ?filename lexbuf parses the lexbuf as an expression and returns the result. This function sets the pos_fname field in the lexbuf to the provided filename. If an error occurs, it returns the error.

parameter filename
The name of the file being parsed.

parameter lexbuf
The lexbuf to be parsed.

returns
The result of the parsing as result.