Module Repl.Execute

Execute

author Bruno Braga

This module provides the implementation of the execute command in the REPL (Read-Eval-Print Loop) of the interpreter. The execute command is responsible for executing a specific event in the program and propagating its effects. It also handles errors that may occur during execution, such as invalid event IDs or type mismatches. The module includes functions to initialize the program, execute events, and handle command-line arguments.

val term : (State.runtime_state -> (State.runtime_state, Ast.Error.detailed_error list) Stdlib.result) Cmdliner.Term.t

term is a command-line term for the execute command.

returns
the term handling the execute command.