Module Typing.Errors

val type_mismatch : ?errors:Ast.Error.detailed_error list -> ?loc:Ast.Syntax.loc -> Ast.Syntax.type_expr' list -> Ast.Syntax.type_expr' list -> ('a, Ast.Error.detailed_error list) Stdlib.result

type_mismatch ?errors ?loc expected actual returns a Result with an error if the expected and actual types do not match.

parameter ?errors
A list of errors to append to the result.

parameter ?loc
The location of the error.

parameter expected
The expected type.

parameter actual
The actual type.

returns
A Result with an error if the expected and actual types do not match.

val event_type_mismatch : ?errors:Ast.Error.detailed_error list -> ?loc:Ast.Syntax.loc -> ?available:(string * (Ast.Syntax.type_expr' * Ast.Syntax.event_type')) list -> (string * Ast.Syntax.event_type' * Ast.Syntax.type_expr') list -> (string * Ast.Syntax.event_type' * Ast.Syntax.type_expr') list -> ('a, Ast.Error.detailed_error list) Stdlib.result

val missing_label : ?errors:Ast.Error.detailed_error list -> ?available_labels:string Ast.Syntax.annotated list -> string Ast.Syntax.annotated -> ('a, Ast.Error.detailed_error list) Stdlib.result

val property_not_found_type : ?errors:Ast.Error.detailed_error list -> ?loc:Ast.Syntax.loc -> string Ast.Syntax.annotated -> Ast.Syntax.type_expr' -> ('a, Ast.Error.detailed_error list) Stdlib.result

val missing_exported_event_types : ?errors:Ast.Error.detailed_error list -> expected:Ast.Syntax.event_id list -> Ast.Syntax.event_id list -> ('a, Ast.Error.detailed_error list) Stdlib.result

val missing_exported_events : ?errors:Ast.Error.detailed_error list -> expected:Ast.Syntax.event_id list -> Ast.Syntax.event_id list -> ('a, Ast.Error.detailed_error list) Stdlib.result